Add ModifiedPipelineDialog
The ModifiedPipelineDialog is used to resolve (potential) conflicts
with modified pipelines that are assigned to/used by at least one
page. The page's pipeline can be updated with or without resetting
it first (thereby deleting all data), pipeline patches can be removed,
or not, and finally, the unmodified pipeline can be restored, turning
the modified pipeline into a clone, and leaving already processed
data intact.