Introduce OSGi bundle fragments
We could do it all with ExtraSystemPackages, but it's nicer to do this
with bundle fragments, where we don't need the convenience of exploring
all packages below some namespace (by means of reflection).
Unfortunately packages in bundle fragments seem to need to be explicitly
listed/enumerated; using a wildcard for the Export-Package instruction
doesn't work.