Fix RgbPage/EditorPresenter.moveToPos bug
While there was/is nothing wrong with the moveToPos method, the
call to it to initially position the page in the editor failed,
due to RgbPage trying to being super clever in running the FX Image
conversion on the thread pool/a background thread, and then later
providing the layer on the FX application thread.

The problem in doing so is that at the point we want to moveToPos
this background thread is still running, an the ZoomPaneSimple of
the editor is still empty. Thus we're running this initial image
conversion on the FX application thread (no problem), and things
work fine again.

Alternatively we could insist on the background job, and introduce
some property we could listen to in the EditorPresenter, but eh...