Move stateProperty from Runnable- to PrototypeProcessor
This moves the stateProperty to the PrototypeProcessor s.t. we
may listen to processor's states in the pipeline editor already.
Since that's a thing now, the glyph color of a processor view now
indicates the state, s.t. users immediately see when a processor
is fully connected.

Secondly the expanded menu is also shown for processors without
parameters, s.t. the processor information dialog is also available
for these processors.