Implement proper pipeline executor
Big patch that primarily implements a proper pipeline executor
with two implementations (a serial and a parallel executor),
and basic pipeline execution loggers (to take execution
timings per pipeline/processor, or for verbose stdout status
updates).
Secondly, execution of individual processors or pipelines may
be cancelled now, although this needs to be supported by the
individual processors (which need to listen for interruption).
Thirdly, several problems with concurrency/synchronization got
fixed.
And finally, this patch also fixes a huge memory leak (of
unreleased pipelines/processors due to dangling references)
while only a single runnable pipeline is supposed to be in
memory at a given point in time.